+++ sentry-rails/spec/sentry/generator_spec.rb
@@ -55,7 +55,7 @@ RSpec.describe SentryGenerator do
       Sentry.init do |config|
         config.breadcrumbs_logger = [:active_support_logger]
         config.dsn = ENV['SENTRY_DSN']
-        config.enable_tracing = true
+        config.traces_sample_rate = 1.0
       end
     RUBY
   end
@@ -87,7 +87,7 @@ RSpec.describe SentryGenerator do
         Sentry.init do |config|
           config.breadcrumbs_logger = [:active_support_logger]
           config.dsn = 'foobarbaz'
-          config.enable_tracing = true
+          config.traces_sample_rate = 1.0
         end
       RUBY
     end